Implement DEBUG file-watch hot-reload and related enhancements
- HUD\_t: include watch\_t; add HUD\_t\$reload\$element callback under DEBUG to reload animation frames and boxes on file changes; register watches in HUD\_t\$load
- background\_t: add background\_t\$reload\$element callback; patch BOXES\_FILE\_EXTENSION usage and animation glob prefix; register watches in background\_t\$load
- watch\_t: change callback signature param name from \_eventId to \_cookie; add error-focused logging in close/add/remove/check paths; ensure watch\_t\$check logs and exits on callback failures
- iterate: add TRY\_CHECK\_WATCH\_OR\_EXIT macro to invoke watch\_t\$check for background watches in DEBUG hot-reload loop
- boxes\_t: define BOXES\_FILE\_EXTENSION macro for consistent “.boxes” usage
- stdfunc: add SECONDS\_TO\_MILLISECONDS macro
- tests: fix duplicateStringTest buffer sizing and use ASSERT\_STRING\_EQ on the copied static buffer
- vsync: include stdfloat16; change desiredFPS type to float16\_t; add guard against re-initialization; cast float16 to float in sleep calculation
- Logging: use minimal, action-focused error/info messages (“Invalid argument”, “Adding watch to path”, etc.)
- Maintain existing behavior outside DEBUG; watch-based hot-reload only active when DEBUG is defined.

+++ vsync/include/vsync.h
@@ -4,6 +4,7 @@
 #include <stdbool.h>
 #include <stddef.h>
 
+#include "stdfloat16.h"
 #include "stdfunc.h"
 #include "vsync_t.h"
 
@@ -35,7 +36,7 @@ vsync_t$convert$fromString( const char* restrict _string ) {
 }
 
 bool vsync$init( const vsync_t _vsync,
-                 const size_t _desiredFPS,
+                 const float16_t _desiredFPS,
                  SDL_Renderer* _renderer );
 bool vsync$quit( void );
